fix multi users issue

+++ backend/src/services/calorieEntry/manager.ts
@@ -6,31 +6,31 @@ import { UserDoc } from "../../models/db/User/User";
 import { UserManager } from "../user/manager";
 
 export class CalorieEntryManager {
-    async addNewEntry(data: CalorieEntryAttr) {
-        const entry = CalorieEntry.build(data);
+    async addNewEntry(data: CalorieEntryAttr, userId: Types.ObjectId) {
+        const entry = CalorieEntry.build({ ...data, user: userId });
         await entry.save();
         return entry;
     }
 
-    async updateEntry(entryId: Types.ObjectId, data: Partial<CalorieEntryAttr>) {
-        const entry = await CalorieEntry.findOneAndUpdate({ _id: entryId }, data, { new: true });
+    async updateEntry(entryId: Types.ObjectId, data: Partial<CalorieEntryAttr>, userId: Types.ObjectId) {
+        const entry = await CalorieEntry.findOneAndUpdate({ _id: entryId, user: userId }, data, { new: true });
         if (!entry) throw new BadRequestError("Calorie Entry not found");
         return entry;
     }
 
-    async getEntriesByDate(dateInput: Date) {
+    async getEntriesByDate(dateInput: Date, userId: Types.ObjectId) {
         const startDate = moment(dateInput).startOf("day").toDate();
         const endDate = moment(dateInput).endOf("day").toDate();
 
-        return this.getEntriesByDateRange(startDate, endDate);
+        return this.getEntriesByDateRange(startDate, endDate, userId);
     }
 
-    async getEntriesByDateRange(startDate: Date, endDate: Date) {
-        return CalorieEntry.find({ createdAt: { $gte: startDate, $lte: endDate } }).sort("-createdAt");
+    async getEntriesByDateRange(startDate: Date, endDate: Date, userId: Types.ObjectId) {
+        return CalorieEntry.find({ createdAt: { $gte: startDate, $lte: endDate }, user: userId }).sort("-createdAt");
     }
 
-    async deleteEntry(entryId: Types.ObjectId) {
-        const entry = await CalorieEntry.findOneAndDelete({ _id: entryId });
+    async deleteEntry(entryId: Types.ObjectId, userId: Types.ObjectId) {
+        const entry = await CalorieEntry.findOneAndDelete({ _id: entryId, user: userId });
         if (!entry) throw new BadRequestError("Calorie Entry not found");
     }
 
@@ -48,7 +48,7 @@ export class CalorieEntryManager {
         const user = await new UserManager().getCompleteUser(userId);
         const bmr = user.gender === "male" ? this.calculateMaleUserBmr(user) : this.calculateFemaleUserBmr(user);
 
-        const entries = await this.getEntriesByDate(dateInput);
+        const entries = await this.getEntriesByDate(dateInput, userId);
         const totalCalories = entries.reduce((t, r) => t + (r.calories > 0 ? r.calories : 0), 0);
 
         const minsSince12AM = moment(dateInput).diff(moment(dateInput).startOf("day"), "minutes");
